pyproject.toml
[project]
name = "edge-auth-manager-prototype"
version = "0.1.0"
description = "Prototype: lightweight agent + auth manager + message-bus TLS channel + dashboard"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "python-jose[cryptography]>=3.3",
  "cryptography>=42.0",
  "httpx>=0.27",
  "tenacity>=8.2",
  "aiormq>=6.8",
  "sqlalchemy>=2.0",
  "aiosqlite>=0.20",
  "streamlit>=1.34",
  "requests>=2.32",
  "plotly>=5.18",
  "pandas>=2.0",
]
